requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
iniconfig==2.1.0
josepy==2.0.0
MarkupSafe==3.0.2
mozilla-django-oidc==4.0.1
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytest==8.3.5
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
whitenoise==6.9.0
django-ratelimit
asgiref==3.8.1
bcrypt==4.3.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.2.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0
asgiref==3.8.1
bcrypt==4.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.2.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
josepy==2.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mozilla-django-oidc==4.0.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
whitenoise==6.9.0
